chore: 뷰모델 관련 코드 디렉토리 정리
