--- app/src/main/java/umc/cozymate/ui/cozy_home/room/room_recommend/RoomRecommendViewModel.kt
@@ -1,7 +1,0 @@
-package umc.cozymate.ui.cozy_home.room.room_recommend
-
-import androidx.lifecycle.ViewModel
-
-class RoomRecommendViewModel : ViewModel() {
-    // TODO: Implement the ViewModel
-}
