+++ app/src/main/java/umc/cozymate/ui/viewmodel/SearchRoomViewModel.kt
@@ -1,4 +1,4 @@
-package umc.cozymate.ui.cozy_home.roommate.search_roommate
+package umc.cozymate.ui.viewmodel
 
 import android.content.Context
 import android.util.Log
@@ -7,41 +7,41 @@ import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import dagger.hilt.android.lifecycle.HiltViewModel
 import dagger.hilt.android.qualifiers.ApplicationContext
-import umc.cozymate.data.model.response.roommate.SearchRoommateResponse
-import umc.cozymate.data.repository.repository.RoommateRepository
+import umc.cozymate.data.model.response.room.SearchRoomResponse
+import umc.cozymate.data.repository.repository.RoomRepository
 import javax.inject.Inject
 
 @HiltViewModel
-class SearchRoommateViewModel @Inject constructor(
-    private val repo: RoommateRepository,
+class SearchRoomViewModel @Inject constructor(
+    private val repo: RoomRepository,
     @ApplicationContext private val context: Context
 ) : ViewModel() {
     private val TAG = this.javaClass.simpleName
     private val _roomId = MutableLiveData<Int>()
     val roomId: LiveData<Int> get() = _roomId
-    private val _searchRoommateResponse = MutableLiveData<SearchRoommateResponse>()
-    val searchRoommateResponse: LiveData<SearchRoommateResponse> get() = _searchRoommateResponse
+    private val _searchRoomResponse = MutableLiveData<SearchRoomResponse>()
+    val searchRoomResponse: LiveData<SearchRoomResponse> get() = _searchRoomResponse
     private val _isLoading = MutableLiveData<Boolean>()
     val isLoading: LiveData<Boolean> = _isLoading
     private val sharedPreferences = context.getSharedPreferences("app_prefs", Context.MODE_PRIVATE)
     fun getToken(): String? {
         return sharedPreferences.getString("access_token", null)
     }
-    // 룸메이트 검색
-    suspend fun getSearchRoommateList(keyword: String) {
+    // 방 검색
+    suspend fun getSearchRoomList(keyword: String) {
         val token = getToken()
         _isLoading.value = true
         if (token != null && keyword != "") {
             try {
-                val response = repo.searchRoommate(token, keyword)
+                val response = repo.searchRoom(token, keyword)
                 if (response.isSuccessful) {
-                    _searchRoommateResponse.value = response.body()
-                    Log.d(TAG, "룸메이트 검색 조회 성공: ${response.body()!!.result}")
+                    _searchRoomResponse.value = response.body()
+                    Log.d(TAG, "방검색 조회 성공: ${response.body()!!.result}")
                 } else {
-                    Log.d(TAG, "룸메이트 검색 조회 에러 메시지: ${response.errorBody()}")
+                    Log.d(TAG, "방검색 조회 에러 메시지: ${response}")
                 }
             } catch (e: Exception) {
-                Log.d(TAG, "룸메이트 검색 조회 api 요청 실패: ${e}")
+                Log.d(TAG, "방검색 조회 api 요청 실패: ${e}")
             } finally {
                 _isLoading.value = false
             }
